# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TPG Reports API
  version: v1.0.6
  description: |
    API for managing reports, templates, and user authentication for TPG Reports.
    Uses HTTP-only cookies for authentication (not Bearer tokens).
    Provides endpoints for creating, reading, updating, and deleting reports, departments, users, and templates, 
    as well as managing the board reporting workflow including section submissions, report finalization, and PDF export.
    
    Authentication: All authenticated endpoints require HTTP-only cookies set by the /login endpoint.
    Frontend requests must include { credentials: 'include' } option.
servers:
  - url: http://localhost:3001/api
    description: Local development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: |
        HTTP-only JWT cookie set by the /login endpoint.
        Frontend must use { credentials: 'include' } in fetch requests.
        Cookie contains JWT token with 12-hour expiry.
  schemas:
    UserRoleEnum:
      type: string
      enum: [secretary, department]
      description: Role of the user.
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
        code:
          type: string
          description: An optional error code.
      required:
        - message
    User:
      type: object
      properties:
        username:
          type: string
          description: Unique username for the user.
          readOnly: true # Typically username is not changed after creation via this DTO
        id:
          type: string
          format: cuid
          description: Unique identifier for the user.
          readOnly: true
        name:
          type: string
          description: Name of the user.
        email:
          type: string
          format: email
          description: Email address of the user (unique).
        role:
          $ref: '#/components/schemas/UserRoleEnum'
        departmentId:
          type: string
          format: cuid
          nullable: true
          description: ID of the department the user belongs to (if applicable).
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the user was created.
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of when the user was last updated.
          readOnly: true
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address.
        password:
          type: string
          format: password
          description: User's password.
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT authentication token.
      required:
        - user
        - token
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: cuid
          readOnly: true
        title:
          type: string
        description:
          type: string
          nullable: true
        cycle:
          type: string
          enum: [WEEKLY, MONTHLY, ADHOC]
        dueAt:
          type: string
          format: date-time
          nullable: true
        state:
          type: string
          enum: [DRAFT, FINAL, PUBLISHED]
          default: DRAFT
        finalizedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        isDeleted:
          type: boolean
          default: false
        templateId:
          type: string
          format: cuid
          nullable: true
          description: ID of the TemplatePack used as a blueprint to create this report, if any.
        createdById:
          type: string
          format: cuid
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ReportSection' # Assuming ReportSection will be fully defined
          readOnly: true
      required:
        - id
        - title
        - cycle
        - state
        - createdById
        - createdAt
        - updatedAt
    ReportSection:
      type: object
      properties:
        id:
          type: string
          format: cuid
          readOnly: true
        reportId:
          type: string
          format: cuid
          readOnly: true
        templateId:
          type: string # This refers to SectionTemplate.id
          format: cuid
          readOnly: true
        key:
          type: string
          description: Key from the SectionTemplate (e.g., 'achievements', 'challenges').
          readOnly: true
        displayName:
          type: string
          description: Display name from the SectionTemplate.
          readOnly: true
        departmentId:
          type: string
          format: cuid
          nullable: true
          readOnly: true
          description: Department responsible for this section.
        isActive:
          type: boolean
          default: true
        contentMarkdown:
          type: string
          nullable: true
          description: Markdown content of the section.
        state:
          type: string
          enum: [DRAFT, SUBMITTED]
          default: DRAFT
          description: Current state of the section.
        submittedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        dueAt:
          type: string
          format: date-time
          nullable: true
        locked:
          type: boolean
          default: false
          description: Whether the section is locked for editing.
        displayOrder:
          type: integer
          description: Order in which the section appears in the report.
        updatedById:
          type: string
          format: cuid
          nullable: true
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - reportId
        - templateId
        - key
        - displayName
        - isActive
        - state
        - locked
        - displayOrder
        - updatedAt
...
    CreateReportRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the report.
        cycle:
          type: string
          enum: [WEEKLY, MONTHLY, ADHOC]
          description: Reporting cycle.
        departmentIds:
          type: array
          items:
            type: string
            format: cuid
          description: List of Department IDs to be included in this report for section creation.
        description:
          type: string
          nullable: true
          description: Optional description for the report.
        # templateId: # This could be an optional field if creating from a specific template directly
        #   type: string
        #   format: cuid
        #   nullable: true
      required:
        - title
        - cycle
        - departmentIds
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object # This will be replaced by specific schema like Report, Template, etc.
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
paths:
...
  /login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticates a user and returns a JWT token along with user details.
      requestBody:
        description: User credentials for login.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful. Returns user object and JWT token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid request payload (e.g., missing email or password).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me:
    get:
      tags:
        - Authentication
      summary: Get Current User Profile
      description: Returns the profile of the currently authenticated user.
      security:
        - bearerAuth: [] # Indicates JWT Bearer token is required
      responses:
        '200':
          description: Successfully retrieved user profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                required:
                  - user
        '401':
          description: Unauthorized. Invalid or missing token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /reports:
    get:
      tags:
        - Reports
      summary: List Reports
      description: Retrieves a list of reports, with optional pagination and filtering.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination.
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of reports per page.
          schema:
            type: integer
            default: 10
        - name: state
          in: query
          description: Filter reports by state.
          schema:
            type: string
            enum: [DRAFT, FINAL, PUBLISHED]
        - name: cycle
          in: query
          description: Filter reports by cycle.
          schema:
            type: string
            enum: [WEEKLY, MONTHLY, ADHOC]
        - name: sortBy
          in: query
          description: Field to sort by (e.g., createdAt, title).
          schema:
            type: string
        - name: sortOrder
          in: query
          description: Sort order.
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: search
          in: query
          description: Search term across title or description.
          schema:
            type: string
        - name: departmentId
          in: query
          description: Filter reports by department.
          schema:
            type: string
            format: cuid
        # Add other potential filters like departmentId, createdById, etc.
      responses:
        '200':
          description: A paginated list of reports.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Reports
      summary: Create Report
      description: |
        Creates a new report, typically for ad-hoc reports or those not based on a template pack.
        Requires 'secretary' role.
      security:
        - bearerAuth: []
      requestBody:
        description: Data for the new report.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
      responses:
        '201':
          description: Report created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden. User does not have 'secretary' role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...